finish refactoring

+++ backend/src/images/images.service.ts
@@ -1,9 +1,10 @@
-import {Injectable, NotFoundException} from "@nestjs/common";
+import {BadRequestException, Injectable, NotFoundException} from "@nestjs/common";
 import {PrismaService} from "../prisma.service";
-import {Image} from "./interfaces";
+import {IImage} from "./interfaces";
 import {MulterFile} from "../storage";
 import {join} from "path"
 import {unlink} from "fs/promises";
+import * as fs from "node:fs";
 
 @Injectable()
 export class ImagesService {
@@ -13,35 +14,53 @@ export class ImagesService {
     async createImages(
         files: MulterFile[],
         superheroId: number,
-    ): Promise<Image[]> {
-        return await Promise.all(
-            files.map(async ({path, originalname: originalName}: MulterFile): Promise<Image> => {
-                const image: Image = await this.prisma.images.create({
-                    data: {path, originalName,},
-                });
+    ): Promise<IImage[]> {
+        try {
+            return await Promise.all(
+                files.map(async ({path, originalname: originalName}: MulterFile): Promise<IImage> => {
+                    const image: IImage = await this.prisma.images.create({
+                        data: {path, originalName,},
+                    });
 
-                await this.prisma.image_Heroes.create({
-                    data: {
-                        superheroId,
-                        imageId: image.id,
-                    },
-                });
+                    await this.prisma.image_Heroes.create({
+                        data: {
+                            superheroId,
+                            imageId: image.id,
+                        },
+                    });
 
-                return image;
-            }),
-        );
+                    return image;
+                }),
+            );
+        } catch (error) {
+            console.error('Error creating images:', error);
+            throw new BadRequestException('Failed to create images');
+        }
     }
 
-    async remove(id: number): Promise<Image> {
-        const image = await this.prisma.images.findUnique({
-            where: {id},
-        });
+    async remove(id: number): Promise<IImage> {
+        try {
+            const image = await this.prisma.images.findUnique({
+                where: {id},
+            });
 
-        if (!image) {
-            throw new NotFoundException("Image not found");
+            if (!image) {
+                throw new NotFoundException("Image not found");
+            }
+
+            const filePath = join(__dirname, '..', '..', image.path);
+            if (fs.existsSync(filePath)) {
+                await unlink(filePath);
+            } else {
+                console.warn(`File not found for image ID ${id}: ${filePath}`);
+            }
+            return await this.prisma.images.delete({where: {id}});
+        } catch (error) {
+            if (error instanceof NotFoundException) {
+                throw error;
+            }
+            console.error('Error remove images:', error);
+            throw new BadRequestException("Error while remove image")
         }
-        const filePath = join(__dirname, '..', '..', image.path);
-        await unlink(filePath);
-        return this.prisma.images.delete({where: {id}});
     }
 }
